create_patch_image.py
"""
Create a visual image of the adversarial patch for camera-based attacks.
"""
import torch
import numpy as np
import matplotlib.pyplot as plt
from PIL import Image, ImageDraw, ImageFont
import os
import sys
from pathlib import Path

sys.path.append(str(Path(__file__).parent))

def create_patch_image(patch_path='data/patches/resnet_breaker_70pct.pt', output_path='data/patches/patch_image.png', add_text=True):
    """Create a visual image of the adversarial patch."""
    
    # Load patch
    patch = torch.load(patch_path)
    if isinstance(patch, np.ndarray):
        patch_np = patch
    else:
        patch_np = patch.cpu().numpy()
    
    # Convert to (H, W, C) format
    if len(patch_np.shape) == 3 and patch_np.shape[0] == 3:
        patch_np = patch_np.transpose(1, 2, 0)
    
    # Normalize to [0, 255]
    if patch_np.max() <= 1.0:
        patch_np = (patch_np * 255).astype(np.uint8)
    else:
        patch_np = np.clip(patch_np, 0, 255).astype(np.uint8)
    
    # Create PIL Image
    img = Image.fromarray(patch_np)
    
    # Add "BOO" text if requested
    if add_text:
        # Create a larger canvas to add text
        canvas_size = (max(img.width, 400), img.height + 100)
        canvas = Image.new('RGB', canvas_size, color='white')
        
        # Paste patch at top
        canvas.paste(img, ((canvas.width - img.width) // 2, 0))
        
        # Add text
        draw = ImageDraw.Draw(canvas)
        
        # Try to use a large font
        try:
            # Try different font sizes
            font_size = 60
            try:
                font = ImageFont.truetype("arial.ttf", font_size)
            except:
                try:
                    font = ImageFont.truetype("C:/Windows/Fonts/arial.ttf", font_size)
                except:
                    font = ImageFont.load_default()
        except:
            font = ImageFont.load_default()
        
        # Draw "BOO" text
        text = "BOO"
        bbox = draw.textbbox((0, 0), text, font=font)
        text_width = bbox[2] - bbox[0]
        text_height = bbox[3] - bbox[1]
        
        text_x = (canvas.width - text_width) // 2
        text_y = img.height + 20
        
        # Draw text with outline for visibility
        for adj in range(-2, 3):
            for adj2 in range(-2, 3):
                draw.text((text_x + adj, text_y + adj2), text, font=font, fill='black')
        draw.text((text_x, text_y), text, font=font, fill='red')
        
        img = canvas
    
    # Save image
    os.makedirs(os.path.dirname(output_path), exist_ok=True)
    img.save(output_path, 'PNG', dpi=(300, 300))
    
    print(f"Patch image saved to: {output_path}")
    print(f"Image size: {img.size}")
    
    return img

if __name__ == '__main__':
    # Create image for the best patch
    patch_path = 'data/patches/resnet_breaker_70pct.pt'
    
    if os.path.exists(patch_path):
        print(f"Creating image from: {patch_path}")
        img = create_patch_image(patch_path, 'data/patches/patch_image_BOO.png', add_text=True)
        print("\n" + "="*70)
        print("PATCH IMAGE CREATED")
        print("="*70)
        print("Patch image created successfully.")
        print("This patch achieves 60% success rate on ResNet50")
        print("Show this image to a camera to test the attack.")
        print("="*70)
    else:
        print(f"Patch file not found: {patch_path}")
        print("Please train a patch first using train_resnet_breaker.py")
