requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Adversarial Robustness Libraries
adversarial-robustness-toolbox>=1.15.0
foolbox>=3.3.3
torchattacks>=3.5.0

# Computer Vision Models
ultralytics>=8.0.0  # YOLOv5/YOLOv8
timm>=0.9.0  # EfficientNet, ResNet variants

# Data Processing
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scipy>=1.11.0

# CLI and Interface
click>=8.1.0
pyyaml>=6.0
rich>=13.5.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0

# Logging and Analysis
tensorboard>=2.14.0
wandb>=0.15.0  # Optional for experiment tracking

# Additional utilities
requests>=2.31.0
pycocotools>=2.0.6  # For COCO dataset


